+++ modules/module-mysql/src/replication/BinLogStream.ts
@@ -148,7 +148,7 @@ export class BinLogStream {
           await this.snapshotTable(connection.connection, batch, result.table);
           await promiseConnection.query('COMMIT');
         } catch (e) {
-          await promiseConnection.query('ROLLBACK');
+          await tryRollback(promiseConnection);
           throw e;
         }
       } finally {
@@ -304,7 +304,7 @@ AND table_type = 'BASE TABLE';`,
       logger.info(`Initial replication done`);
       await promiseConnection.query('COMMIT');
     } catch (e) {
-      await promiseConnection.query('ROLLBACK');
+      await tryRollback(promiseConnection);
       throw e;
     } finally {
       connection.release();
@@ -665,5 +665,13 @@ AND table_type = 'BASE TABLE';`,
       default:
         return null;
     }
+  }
+}
+
+async function tryRollback(promiseConnection: mysqlPromise.Connection) {
+  try {
+    await promiseConnection.query('ROLLBACK');
+  } catch (e) {
+    logger.error('Failed to rollback transaction', e);
   }
 }