+++ modules/module-mongodb/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @powersync/service-module-mongodb
 
+## 0.7.2
+
+### Patch Changes
+
+- 535e708: Increase connection timeouts for MongoDB "Test Connection".
+- Updated dependencies [f049f68]
+- Updated dependencies [8601d6c]
+  - @powersync/service-core@1.9.0
+
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [7348ea0]
+- Updated dependencies [7d1cd98]
+  - @powersync/service-core@1.8.1
+  - @powersync/lib-service-mongodb@0.5.1
+
 ## 0.7.0
 
 ### Minor Changes