+++ modules/module-postgres/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @powersync/service-module-postgres
 
+## 0.10.0
+
+### Minor Changes
+
+- f049f68: [Postgres] Only flush once per replicated chunk, increasing transaction replication throughput.
+
+### Patch Changes
+
+- Updated dependencies [f049f68]
+- Updated dependencies [8601d6c]
+  - @powersync/service-core@1.9.0
+
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies [7348ea0]
+  - @powersync/service-core@1.8.1
+
 ## 0.9.0
 
 ### Minor Changes