+++ packages/service-core-tests/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/service-core-tests
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [f049f68]
+- Updated dependencies [8601d6c]
+  - @powersync/service-core@1.9.0
+
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [7348ea0]
+  - @powersync/service-core@1.8.1
+
 ## 0.7.0
 
 ### Minor Changes