+++ test-client/CHANGELOG.md
@@ -1,5 +1,20 @@
 # test-client
 
+## 0.1.31
+
+### Patch Changes
+
+- Updated dependencies [f049f68]
+- Updated dependencies [8601d6c]
+  - @powersync/service-core@1.9.0
+
+## 0.1.30
+
+### Patch Changes
+
+- Updated dependencies [7348ea0]
+  - @powersync/service-core@1.8.1
+
 ## 0.1.29
 
 ### Patch Changes